+++ packages/ens-webfont/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.1",
   "type": "module",
   "files": [
-    "index.ts",
+    "dist",
     "Unifont.otf",
     "font-config.css"
   ],
